Commented out some stuff and uncommented other stuff
Commented out the faulty if/else statement.
Uncommented enemy buffer rendering and the filter that finds enemies
that are big enough to split at the bot.
Plus Atom wanted to fix a bunch of white space, and I took out a couple
of small and absolutely unnecessary things, that my linter pointed out,
like an extra semicolon.