Added persistent destinations
I’ll try to keep this short.

Added two variables outside of `findDestination`. One is an object that
is used to remember a list of destinations to go to between calls to
`findDestination()`. The other is the location of the cluster the bot
wants to go to next.

Moved the code to find the closest cluster/supercluster into a function
to save the bot thinking time, because that doesn’t need to be run if
the bot already has a cluster to go to in mind.

If the bot doesn’t have a destination cluster in mind, it will look for
more clusters. If it finds one(or more in the case of a supercluster),
it stores those in memory. One by one it moves to those clusters. If it
starts running from a bot, it will erase the cluster locations from
memory, because if the bot travels a far distance, those locations are
no longer relevant.